componentize MarkdownEditor component
