+++ src/app/problem/edit/ProblemEditorTitle.css
@@ -28,6 +28,7 @@
     left: 133px;
     top: 79px;
     margin-left: 150px;
+    margin-bottom: 20px;
 
     background: #F9F9F9;
     border: 1px solid #C6C4C4;
